Released movies will apply a genre-penalty buff that will be used for other movies score-calculation

+++ Assets/Scripts/Core/Buffs/Buff.cs
@@ -5,7 +5,7 @@ namespace SameOldStory.Core.Buffs {
     public abstract class Buff {
 
         public string Key { get; protected set;}
-        public float Value { get; protected set;}
+        public int Value { get; protected set;}
         
         protected Buff() {
